Deduplicate self-scope authorization expansion guard
